Refactor App.jsx
